Update README.md
The user picks a choice between rock, paper and scissor and "computer" chooses by uses Math.random() to pick a random number between 1 - 3.
That random number is then associated with a choice between rock, paper, scissors and matched against the Math.random() "computer" choice. 
The results are displayed along with a winner, loser or draw.